Revert "switched to FP64 for scalar flux vectors. This greatly improves testing reproducibility and allows for elimination of tolerant test harness variant."
This reverts commit 15a76f06bcaa558fa5a2183cdae4060778fbb533.

+++ include/openmc/random_ray/flat_source_domain.h
@@ -144,8 +144,8 @@ public:
 
   // 2D arrays stored in 1D representing values for all source regions x energy
   // groups
-  vector<double> scalar_flux_old_;
-  vector<double> scalar_flux_new_;
+  vector<float> scalar_flux_old_;
+  vector<float> scalar_flux_new_;
   vector<float> source_;
   vector<float> external_source_;
   vector<bool> external_source_present_;
